# api/swagger.yaml
basePath: /beetle
definitions:
...
  controller.GenerateSourceInfraSummaryRequest:
    properties:
      onpremiseInfraModel:
        $ref: '#/definitions/onpremisemodel.OnpremInfra'
    required:
    - onpremiseInfraModel
    type: object
...
  model.ApiResponse-any:
    properties:
      data:
        description: Contains the actual response data (single object, list, or page)
      error:
        description: Error message for failed responses
        example: Error message if failure
        type: string
      message:
        description: Optional message for additional context
        example: Operation successful
        type: string
      success:
        description: Indicates whether the API call was successful
        example: true
        type: boolean
    type: object
...
  model.ApiResponse-summary_SourceInfraSummary:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/summary.SourceInfraSummary'
        description: Contains the actual response data (single object, list, or page)
      error:
        description: Error message for failed responses
        example: Error message if failure
        type: string
      message:
        description: Optional message for additional context
        example: Operation successful
        type: string
      success:
        description: Indicates whether the API call was successful
        example: true
        type: boolean
    type: object
  model.ApiResponse-summary_TargetInfraSummary:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/summary.TargetInfraSummary'
        description: Contains the actual response data (single object, list, or page)
      error:
        description: Error message for failed responses
        example: Error message if failure
        type: string
      message:
        description: Optional message for additional context
        example: Operation successful
        type: string
      success:
        description: Indicates whether the API call was successful
        example: true
        type: boolean
    type: object
...
  onpremisemodel.CpuProperty:
    properties:
      architecture:
        example: x86_64
        type: string
      cores:
        description: Number of physical cores per CPU
        example: 18
        type: integer
      cpus:
        description: Number of physical CPUs (sockets)
        example: 2
        type: integer
      maxSpeed:
        description: Maximum speed in GHz
        example: 3.6
        type: number
      model:
        example: Intel(R) Xeon(R) Gold 6140 CPU @ 2.30GHz
        type: string
      threads:
        description: Number of logical CPUs (threads) per CPU with hyper-threading
          enabled
        example: 36
        type: integer
      vendor:
        example: GenuineIntel
        type: string
    required:
    - cores
    - cpus
    - threads
    type: object
  onpremisemodel.DiskProperty:
    properties:
      available:
        description: Unit GiB
        type: integer
      label:
        type: string
      totalSize:
        description: Unit GiB
        example: 1024
        type: integer
      type:
        description: SSD, HDD
        example: SSD
        type: string
      used:
        description: Unit GiB
        type: integer
    required:
    - label
    - totalSize
    - type
    type: object
  onpremisemodel.FirewallRuleProperty:
    properties:
      action:
        description: e.g., allow, deny
        type: string
      direction:
        description: e.g., inbound, outbound
        type: string
      dstCIDR:
        description: e.g., "123.123.123.123/32", "123.123.123.123/24", "0.0.0.0/0"
        type: string
      dstPorts:
        description: e.g., "80", "80,443", "1024-65535", "*" (for all ports)
        type: string
      protocol:
        description: e.g., "TCP", "UDP", "ICMP", "*" (for all protocol)
        type: string
      srcCIDR:
        description: e.g., "123.123.123.123/32", "123.123.123.123/24", "0.0.0.0/0"
        type: string
      srcPorts:
        description: e.g., "80", "80,443", "1024-65535", "*" (for all ports)
        type: string
    type: object
...
  onpremisemodel.MemoryProperty:
    properties:
      available:
        description: Unit GiB
        type: integer
      totalSize:
        description: Unit GiB
        example: 128
        type: integer
      type:
        example: DDR4
        type: string
      used:
        description: Unit GiB
        type: integer
    required:
    - totalSize
    - type
    type: object
  onpremisemodel.NetworkDetail:
    properties:
      cidrBlocks:
        items:
          type: string
        type: array
      defaultGateways:
        items:
          $ref: '#/definitions/onpremisemodel.GatewayProperty'
        type: array
    type: object
  onpremisemodel.NetworkInterfaceProperty:
    properties:
      ipv4CidrBlocks:
        description: IPv4 address with prefix length (e.g., 192.168.0.21/24), instead
          of inet addr, Bcast, and Mask
        items:
          type: string
        type: array
      ipv6CidrBlocks:
        description: IPv6 address with prefix length (e.g., "2001:db8::1/64")
        items:
          type: string
        type: array
      macAddress:
        description: MAC address
        type: string
      mtu:
        description: Maximum Transmission Unit (MTU) in bytes
        type: integer
      name:
        description: Interface name (e.g., eth0, ens01, enp0s3)
        type: string
      state:
        description: Interface state (e.g., UP, DOWN)
        type: string
    required:
    - name
    type: object
  onpremisemodel.NetworkProperty:
    properties:
      ipv4Networks:
        $ref: '#/definitions/onpremisemodel.NetworkDetail'
      ipv6Networks:
        allOf:
        - $ref: '#/definitions/onpremisemodel.NetworkDetail'
        description: TBD
    type: object
  onpremisemodel.OnpremInfra:
    properties:
      network:
        $ref: '#/definitions/onpremisemodel.NetworkProperty'
      servers:
        items:
          $ref: '#/definitions/onpremisemodel.ServerProperty'
        type: array
    required:
    - servers
    type: object
  onpremisemodel.OsProperty:
    properties:
      id:
        example: ubuntu
        type: string
      idLike:
        example: debian
        type: string
      name:
        example: Ubuntu
        type: string
      prettyName:
        description: Pretty name
        example: Ubuntu 22.04.3 LTS
        type: string
      version:
        description: Full version string
        example: 22.04.3 LTS (Jammy Jellyfish)
        type: string
      versionCodename:
        example: jammy
        type: string
      versionId:
        example: "22.04"
        type: string
    required:
    - prettyName
    type: object
  onpremisemodel.RouteProperty:
    properties:
      destination:
        description: Destination network, expressed in CIDR format
        type: string
      gateway:
        description: Gateway address to which packets are forwarded
        type: string
      interface:
        description: Network interface associated with the route
        type: string
      linkState:
        description: Link state of the route (e.g., UP, DOWN)
        type: string
      metric:
        description: Metric value indicating the priority of the route
        type: integer
      protocol:
        description: Protocol used to set the route (e.g., kernel, static)
        type: string
      scope:
        description: Scope of the route (e.g., global, link, host)
        type: string
      source:
        description: Optionally stores the source address (used for policy-based routing)
        type: string
    type: object
  onpremisemodel.ServerProperty:
    properties:
      cpu:
        $ref: '#/definitions/onpremisemodel.CpuProperty'
      dataDisks:
        items:
          $ref: '#/definitions/onpremisemodel.DiskProperty'
        type: array
      firewallTable:
        items:
          $ref: '#/definitions/onpremisemodel.FirewallRuleProperty'
        type: array
      hostname:
        type: string
      interfaces:
        items:
          $ref: '#/definitions/onpremisemodel.NetworkInterfaceProperty'
        type: array
      machineId:
        description: Unique identifier for the server (e.g., UUID)
        type: string
      memory:
        $ref: '#/definitions/onpremisemodel.MemoryProperty'
      os:
        $ref: '#/definitions/onpremisemodel.OsProperty'
      rootDisk:
        $ref: '#/definitions/onpremisemodel.DiskProperty'
      routingTable:
        items:
          $ref: '#/definitions/onpremisemodel.RouteProperty'
        type: array
    type: object
...
  summary.SourceInfraSummary:
    properties:
      computeResources:
        $ref: '#/definitions/summary.SourceSummaryComputeResources'
      networkResources:
        $ref: '#/definitions/summary.SourceSummaryNetworkResources'
      overview:
        $ref: '#/definitions/summary.SourceSummaryInfraOverview'
      securityResources:
        $ref: '#/definitions/summary.SourceSummarySecurityResources'
      storageResources:
        $ref: '#/definitions/summary.SourceSummaryStorageResources'
      summaryMetadata:
        $ref: '#/definitions/summary.SourceSummaryMetadata'
    type: object
...
  summary.SourceNetworkInfo:
    properties:
      description:
        example: Office LAN network
        type: string
      dnsServers:
        example:
        - 8.8.8.8
        - 8.8.4.4
        items:
          type: string
        type: array
      gateway:
        example: 192.168.1.1
        type: string
      networkCidr:
        example: 192.168.1.0/24
        type: string
      networkName:
        example: office-lan
        type: string
      serverCount:
        example: 5
        type: integer
    type: object
...
  summary.SourceServerFirewall:
    properties:
      firewallRules:
        items:
          $ref: '#/definitions/summary.SourceFirewallRule'
        type: array
      hostname:
        example: web-server-01
        type: string
    type: object
  summary.SourceServerInfo:
    properties:
      cpu:
        $ref: '#/definitions/summary.SourceCPUInfo'
      disk:
        $ref: '#/definitions/summary.SourceDiskInfo'
      hostname:
        example: web-server-01
        type: string
      interfaces:
        items:
          $ref: '#/definitions/summary.SourceInterfaceInfo'
        type: array
      machineId:
        example: 00a9f3d4-74b6-e811-906e-000ffee02d5c
        type: string
      memory:
        $ref: '#/definitions/summary.SourceMemoryInfo'
      network:
        $ref: '#/definitions/summary.SourceNetworkBrief'
      os:
        $ref: '#/definitions/summary.SourceOSInfo'
      rootDiskSize:
        example: 100
        type: integer
      rootDiskType:
        example: SSD
        type: string
      routingTable:
        items:
          $ref: '#/definitions/summary.SourceRoutingTableRow'
        type: array
    type: object
  summary.SourceStorageByServer:
    properties:
      hostname:
        example: web-server-01
        type: string
      totalGb:
        example: 500
        type: integer
      type:
        example: SSD
        type: string
    type: object
  summary.SourceStorageByType:
    properties:
      serverCount:
        example: 3
        type: integer
      totalGb:
        example: 1000
        type: integer
      type:
        example: SSD
        type: string
    type: object
  summary.SourceSummaryComputeResources:
    properties:
      servers:
        items:
          $ref: '#/definitions/summary.SourceServerInfo'
        type: array
    type: object
  summary.SourceSummaryInfraOverview:
    properties:
      infraName:
        example: datacenter-seoul-1
        type: string
      totalCpuCores:
        example: 40
        type: integer
      totalDiskGb:
        example: 2000
        type: integer
      totalMemoryGb:
        example: 160
        type: integer
      totalNetworks:
        example: 3
        type: integer
      totalServerCount:
        example: 5
        type: integer
    type: object
  summary.SourceSummaryMetadata:
    properties:
      generatedAt:
        example: "2025-10-31T14:30:00Z"
        type: string
      infraName:
        example: datacenter-seoul-1
        type: string
      summaryVersion:
        example: "1.0"
        type: string
    type: object
  summary.SourceSummaryNetworkResources:
    properties:
      networks:
        items:
          $ref: '#/definitions/summary.SourceNetworkInfo'
        type: array
    type: object
  summary.SourceSummarySecurityResources:
    properties:
      serverFirewalls:
        items:
          $ref: '#/definitions/summary.SourceServerFirewall'
        type: array
    type: object
  summary.SourceSummaryStorageResources:
    properties:
      availableGb:
        example: 800
        type: integer
      byServer:
        items:
          $ref: '#/definitions/summary.SourceStorageByServer'
        type: array
      byType:
        items:
          $ref: '#/definitions/summary.SourceStorageByType'
        type: array
      totalGb:
        example: 2000
        type: integer
      usedGb:
        example: 1200
        type: integer
    type: object
  summary.SummaryComputeResources:
    properties:
      images:
        items:
          $ref: '#/definitions/summary.SummaryImageInfoWithUsage'
        type: array
      specs:
        items:
          $ref: '#/definitions/summary.SummarySpecInfoWithUsage'
        type: array
      vms:
        items:
          $ref: '#/definitions/summary.SummaryVmInfo'
        type: array
    type: object
  summary.SummaryCostByRegion:
    properties:
      costPerHour:
        example: 0.4512
        type: number
      costPerMonth:
        example: 324.86
        type: number
      csp:
        example: AWS
        type: string
      region:
        example: ap-northeast-2
        type: string
      vmCount:
        example: 3
        type: integer
    type: object
...
  summary.SummaryCostEstimation:
    properties:
      byRegion:
        items:
          $ref: '#/definitions/summary.SummaryCostByRegion'
        type: array
      byVm:
        items:
          $ref: '#/definitions/summary.SummaryCostByVm'
        type: array
      currency:
        example: USD
        type: string
      totalCostPerDay:
        example: 10.83
        type: number
      totalCostPerHour:
        example: 0.4512
        type: number
      totalCostPerMonth:
        example: 324.86
        type: number
    type: object
...
  summary.SummaryNetworkResources:
    properties:
      vnets:
        items:
          $ref: '#/definitions/summary.SummaryVNetInfo'
        type: array
    type: object
...
  summary.SummarySecurityResources:
    properties:
      securityGroups:
        items:
          $ref: '#/definitions/summary.SummarySecurityGroupInfo'
        type: array
      sshKeys:
        items:
          $ref: '#/definitions/summary.SummarySshKeyInfo'
        type: array
    type: object
...
  summary.TargetInfraOverview:
    properties:
      installMonAgent:
        example: "no"
        type: string
      label:
        additionalProperties:
          type: string
        type: object
      mciDescription:
        example: Migrated infrastructure
        type: string
      mciName:
        example: mmci01
        type: string
      runningVmCount:
        example: 3
        type: integer
      status:
        example: Running
        type: string
      stoppedVmCount:
        example: 0
        type: integer
      targetCloud:
        example: AWS
        type: string
      targetRegion:
        example: ap-northeast-2
        type: string
      totalVmCount:
        example: 3
        type: integer
    type: object
  summary.TargetInfraSummary:
    properties:
      computeResources:
        $ref: '#/definitions/summary.SummaryComputeResources'
      costEstimation:
        $ref: '#/definitions/summary.SummaryCostEstimation'
      networkResources:
        $ref: '#/definitions/summary.SummaryNetworkResources'
      overview:
        $ref: '#/definitions/summary.TargetInfraOverview'
      securityResources:
        $ref: '#/definitions/summary.SummarySecurityResources'
      summaryMetadata:
        $ref: '#/definitions/summary.TargetSummaryMetadata'
    type: object
  summary.TargetSummaryMetadata:
    properties:
      generatedAt:
        example: "2025-10-31T14:30:00Z"
        type: string
      mciId:
        example: mmci01
        type: string
      mciName:
        example: my-migrated-infrastructure
        type: string
      namespace:
        example: mig01
        type: string
      summaryVersion:
        example: "1.0"
        type: string
    type: object
...
host: localhost:8056
info:
  contact:
    name: API Support
    url: https://github.com/cloud-barista/cm-beetle/issues/new/choose
  description: CM-Beetle recommends optimal target cloud infrastructure and executes
    migration.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: CM-Beetle API
  version: latest
paths:
...
  /summary/source:
    post:
      consumes:
      - application/json
      description: |-
        Generate a comprehensive source infrastructure summary from on-premise data in multiple formats based on 'format' query parameter:

        **Response Format by 'format' Parameter:**
        - `format=json`: Returns ApiResponse[SourceInfraSummary] with Content-Type: application/json
        - `format=md` (default): Returns markdown string with Content-Type: text/markdown; charset=utf-8
        - `format=html`: Returns HTML string with Content-Type: text/html; charset=utf-8

        **Note:** API documentation shows JSON schema for reference, but actual default response is markdown format.

        **Markdown example**: https://github.com/cloud-barista/cm-beetle/blob/main/cmd/test-cli/testresult/beetle-summary-source.md

        **Download Behavior:**
        - `download=false` (default): Content displayed inline (viewable in browser/Swagger UI)
        - `download=true`: Content downloaded as file (Content-Disposition: attachment)
      operationId: GenerateSourceInfraSummary
      parameters:
      - default: md
        description: 'Summary format: md, html, or json'
        enum:
        - md
        - html
        - json
        in: query
        name: format
        type: string
      - default: "false"
        description: 'Download as file: true for file download, false for inline display
          (only affects browsers/Swagger UI, not curl)'
        enum:
        - "true"
        - "false"
        in: query
        name: download
        type: string
      - description: Unique request ID (auto-generated if not provided). Used for
          tracking request status and correlating logs.
        in: header
        name: X-Request-Id
        type: string
      - description: Source infrastructure data
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/controller.GenerateSourceInfraSummaryRequest'
      produces:
      - application/json
      - text/markdown
      - text/html
      responses:
        "200":
          description: Successfully generated source infrastructure summary (format
            varies by 'format' parameter)
          headers:
            Content-Disposition:
              description: inline; filename=\"source-summary.md\" or \"source-summary.html\"
                (or attachment when download=true)
              type: string
            Content-Type:
              description: text/markdown; charset=utf-8 or text/html; charset=utf-8
              type: string
          schema:
            $ref: '#/definitions/model.ApiResponse-summary_SourceInfraSummary'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/model.ApiResponse-any'
        "500":
          description: Internal server error during summary generation
          schema:
            $ref: '#/definitions/model.ApiResponse-any'
      summary: Generate source infrastructure summary
      tags:
      - '[Summary/Report] Infrastructure Analysis for Migration'
  /summary/target/ns/{nsId}/mci/{mciId}:
    get:
      consumes:
      - application/json
      description: |-
        Generate a comprehensive target infrastructure summary in multiple formats based on 'format' query parameter:

        **Response Format by 'format' Parameter:**
        - `format=md` (default): Returns markdown string with Content-Type: text/markdown; charset=utf-8
        - `format=html`: Returns HTML string with Content-Type: text/html; charset=utf-8
        - `format=json`: Returns ApiResponse[TargetInfraSummary] with Content-Type: application/json

        **Note:** API documentation shows JSON schema for reference, but actual default response is markdown format.

        **Markdown example**: https://github.com/cloud-barista/cm-beetle/blob/main/cmd/test-cli/testresult/beetle-summary-target-aws.md

        **Download Behavior:**
        - `download=false` (default): Content displayed inline (viewable in browser/Swagger UI)
        - `download=true`: Content downloaded as file (Content-Disposition: attachment)
      operationId: GenerateTargetInfraSummary
      parameters:
      - default: mig01
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - default: mmci01
        description: Multi-Cloud Infrastructure (MCI) ID
        in: path
        name: mciId
        required: true
        type: string
      - default: md
        description: 'Summary format: md, html, or json'
        enum:
        - md
        - html
        - json
        in: query
        name: format
        type: string
      - default: "false"
        description: 'Download as file: true for file download, false for inline display
          (only affects browsers/Swagger UI, not curl)'
        enum:
        - "true"
        - "false"
        in: query
        name: download
        type: string
      - description: Unique request ID (auto-generated if not provided). Used for
          tracking request status and correlating logs.
        in: header
        name: X-Request-Id
        type: string
      produces:
      - application/json
      - text/markdown
      - text/html
      responses:
        "200":
          description: Successfully generated target infrastructure summary (format
            varies by 'format' parameter)
          headers:
            Content-Disposition:
              description: inline; filename=\"target-summary.md\" or \"target-summary.html\"
                (or attachment when download=true)
              type: string
            Content-Type:
              description: text/markdown; charset=utf-8 or text/html; charset=utf-8
              type: string
          schema:
            $ref: '#/definitions/model.ApiResponse-summary_TargetInfraSummary'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/model.ApiResponse-any'
        "500":
          description: Internal server error during summary generation
          schema:
            $ref: '#/definitions/model.ApiResponse-any'
      summary: Generate target infrastructure summary
      tags:
      - '[Summary/Report] Infrastructure Analysis for Migration'
...
security:
- BasicAuth: []
- BearerAuth: []
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: Enter your bearer token in the format **Bearer &lt;token&gt;**
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...